nn-syntax: add the ability to override how entities are allocated
Going forward, we'll need different rules for how entities are
matched to the sentence, and potentially different dataset-specific
HACKS.

The right place for that is in Genie, which knows what a sentence
is in the first place, and can tokenize and stuff.
